+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - Schema: add `isMassAssignable` attribute to fields
 - Sequelize: honour `existsInClient` and `existsInServer` in `model.toJSON()` and `isMassAssignable` in `model.set()`
+- GraphQL: don't include models/fields if either `existsInServer` or `existsInClient` are unset.
 
 ## 0.3.0 (2018-4-27)